separated trt 7 ops
