+++ torch2trt/converters/trt7_ops/avg_pool.py
